requirements.txt
playwright>=1.40.0
requests>=2.28.2
selenium>=4.7.2
python-dotenv>=1.0.0
apit212>=3.0.1
pydantic>=2.0.0
playwright-stealth>=1.0.1
pandas>=1.3.0
lxml>=4.9.0
aiohttp>=3.9.0
yarl>=1.9.0
